refactor: App클래스로부터 Result클래스로 분리
